// src/modules/analytic/models/responses/ReportConversionStaffResponse.ts
export interface ReportConversionStaffResponse {
  assigneeName: string;
  trafficAssigneeQuantity: number;
  customerPurchase: number;
  crAssigneePurchase: number;
  trafficNotBoughtQuantity: number;
  trafficAssigneeQuantityGrowthRatio: number;
  customerPurchaseGrowthRatio: number;
  crAssigneePurchaseGrowthRatio: number;
  trafficNotBoughtQuantityGrowthRatio: number;
}